Revert "feat: add list component and use it in glucose-notification-list"
This reverts commit 2989de81a8c6093036d6572f3e7804f129c298c0.
